+++ cli/cli/scripts/build.sh
@@ -89,7 +89,16 @@ fi
 # Final verification
 goarch="$(go env GOARCH)"
 goos="$(go env GOOS)"
-cli_binary_filepath="${cli_module_dirpath}/${GORELEASER_OUTPUT_DIRNAME}/${GORELEASER_CLI_BUILD_ID}_${goos}_${goarch}/${CLI_BINARY_FILENAME}"
+architurecture_dirname="${GORELEASER_CLI_BUILD_ID}_${goos}_${goarch}"
+if [ "${goarch}" == "${GO_ARCH_ENV_AMD64_VALUE}" ]; then
+  goamd64="$(go env GOAMD64)"
+  if [ "${goamd64}" == "" ]; then
+    goamd64="${GO_DEFAULT_AMD64_ENV}"
+  fi
+  architurecture_dirname="${architurecture_dirname}_${goamd64}"
+fi
+
+cli_binary_filepath="${cli_module_dirpath}/${GORELEASER_OUTPUT_DIRNAME}/${architurecture_dirname}/${CLI_BINARY_FILENAME}"
 if ! [ -f "${cli_binary_filepath}" ]; then
     echo "Error: Expected a CLI binary to have been built by Goreleaser at '${cli_binary_filepath}' but none exists" >&2
     exit 1